pyproject.toml
[tool.poetry]
name = "frogbot-race-repro"
version = "0.1.0"
description = "Reproduce Frogbot CA race condition"
authors = ["test <test@test.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "5.3.1"
jinja2 = "3.0.0"
werkzeug = "1.0.0"
flask = "1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
